getting scriptureforge e2e tests passing after merge
Former-commit-id: 8c67a155d4672f204971e9ab32f8e9b2c880ae0f

+++ test/app/teardownTestEnvironment.php
@@ -19,12 +19,15 @@ if ($constants['siteType'] == 'languageforge') {
 
 function recursiveRemoveFolder($folder)
 {
-    foreach (glob("{$folder}/*") as $file) {
-        if (is_dir($file)) {
-            recursiveRemoveFolder($file);
-        } else {
-            unlink($file);
-        }
-    }
-    rmdir($folder);
+	if (is_dir($folder)) {
+	    foreach (glob("{$folder}/*") as $file) {
+	        if (is_dir($file)) {
+	            recursiveRemoveFolder($file);
+	        } else {
+	        	if (is_file($file)) {
+	            	unlink($file);
+	        	}
+	        }
+	    }
+	}
 }
